implemented a simplified example.
