+++ example/pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   cupertino_icons: ^0.1.3
   path_provider: ^1.6.1
   path: ^1.6.4
+  permission_handler: ">=5.0.0 <6.0.0"
   uuid: ^2.0.4
 
 dev_dependencies:
